move date_time_input to a more shared location not scoped to custom status
