+++ webapp/channels/src/components/date_time_picker_modal/date_time_picker_modal.tsx
@@ -10,7 +10,7 @@ import {GenericModal} from '@mattermost/components';
 
 import {getCurrentTimezone} from 'mattermost-redux/selectors/entities/timezone';
 
-import DateTimeInput, {getRoundedTime} from 'components/custom_status/date_time_input';
+import DateTimeInput, {getRoundedTime} from 'components/datetime_input/datetime_input';
 
 import Constants from 'utils/constants';
 import {isKeyPressed} from 'utils/keyboard';
