+++ webapp/channels/src/components/post_reminder_custom_time_picker_modal/post_reminder_custom_time_picker_modal.tsx
@@ -5,8 +5,8 @@ import type {Moment} from 'moment-timezone';
 import React, {useCallback} from 'react';
 import {useIntl} from 'react-intl';
 
-import {getRoundedTime} from 'components/custom_status/date_time_input';
 import DateTimePickerModal from 'components/date_time_picker_modal/date_time_picker_modal';
+import {getRoundedTime} from 'components/datetime_input/datetime_input';
 
 import {toUTCUnixInSeconds} from 'utils/datetime';
 import {getCurrentMomentForTimezone} from 'utils/timezone';
