+++ webapp/channels/src/i18n/en.json
@@ -3887,10 +3887,11 @@
   "discard_changes_modal.leave": "Yes, Discard",
   "discard_changes_modal.message": "You have unsaved changes, are you sure you want to discard them?",
   "discard_changes_modal.title": "Discard Changes?",
+  "datetime_picker.date": "Date",
+  "datetime_picker.time": "Time",
   "dnd_custom_time_picker_modal.date": "Date",
   "dnd_custom_time_picker_modal.defaultMsg": "Disable notifications until",
   "dnd_custom_time_picker_modal.submitButton": "Disable Notifications",
-  "dnd_custom_time_picker_modal.time": "Time",
   "drafts.actions.delete": "Delete draft",
   "drafts.actions.edit": "Edit draft",
   "drafts.actions.scheduled": "Schedule draft",
